double -> float

+++ Sources/ImageThumbnailer/Lib.swift
@@ -16,11 +16,11 @@ public struct ThumbnailInfo {
 }
 
 public struct GPSLocation {
-    public let latitude: Double
-    public let longitude: Double
-    public let altitude: Double
+    public let latitude: Float
+    public let longitude: Float
+    public let altitude: Float
 
-    public init(latitude: Double, longitude: Double, altitude: Double) {
+    public init(latitude: Float, longitude: Float, altitude: Float) {
         self.latitude = latitude
         self.longitude = longitude
         self.altitude = altitude
@@ -30,11 +30,11 @@ public struct GPSLocation {
 public struct Metadata {
     public let width: UInt32
     public let height: UInt32
-    public let duration: Double?  // Duration in seconds
+    public let duration: Float?  // Duration in seconds
     public let location: GPSLocation?
 
     public init(
-        width: UInt32, height: UInt32, duration: Double? = nil, location: GPSLocation? = nil
+        width: UInt32, height: UInt32, duration: Float? = nil, location: GPSLocation? = nil
     ) {
         self.width = width
         self.height = height
